+++ src/Views/Filters/MultiSelectDropdownFilter.php
@@ -3,48 +3,15 @@
 namespace Rappasoft\LaravelLivewireTables\Views\Filters;
 
 use Rappasoft\LaravelLivewireTables\Views\Filter;
+use Rappasoft\LaravelLivewireTables\Views\Traits\Filters\{HasOptions,IsArrayFilter};
 
 class MultiSelectDropdownFilter extends Filter
 {
-    public array $options = [];
+    use HasOptions,
+        IsArrayFilter;
 
     protected string $view = 'livewire-tables::components.tools.filters.multi-select-dropdown';
 
-    protected string $firstOption = '';
-
-    public function setFirstOption(string $firstOption): MultiSelectDropdownFilter
-    {
-        $this->firstOption = $firstOption;
-
-        return $this;
-    }
-
-    public function getFirstOption(): string
-    {
-        return $this->firstOption;
-    }
-
-    public function options(array $options = []): MultiSelectDropdownFilter
-    {
-        $this->options = $options;
-
-        return $this;
-    }
-
-    public function getOptions(): array
-    {
-        return $this->options;
-    }
-
-    public function getKeys(): array
-    {
-        return collect($this->getOptions())
-            ->keys()
-            ->map(fn ($value) => (string) $value)
-            ->filter(fn ($value) => strlen($value))
-            ->values()
-            ->toArray();
-    }
 
     public function validate(int|string|array $value): array|int|string|bool
     {
@@ -62,21 +29,6 @@ class MultiSelectDropdownFilter extends Filter
         return (is_string($value) || is_numeric($value)) ? $value : false;
     }
 
-    /**
-     * Get the filter default options.
-     */
-    public function getDefaultValue(): array
-    {
-        return [];
-    }
-
-    /**
-     * Gets the Default Value for this Filter via the Component
-     */
-    public function getFilterDefaultValue(): array
-    {
-        return $this->filterDefaultValue ?? [];
-    }
 
     public function getFilterPillValue($value): ?string
     {
