+++ src/Views/Traits/Filters/HasOptions.php
@@ -1,0 +1,49 @@
+<?php
+
+namespace Rappasoft\LaravelLivewireTables\Views\Traits\Filters;
+
+use Closure;
+use Illuminate\View\ComponentAttributeBag;
+use Rappasoft\LaravelLivewireTables\Views\{Column,Filter};
+
+trait HasOptions
+{
+    public array $options = [];
+
+    protected string $firstOption = '';
+
+    public function setFirstOption(string $firstOption): self
+    {
+        $this->firstOption = $firstOption;
+
+        return $this;
+    }
+
+    public function getFirstOption(): string
+    {
+        return $this->firstOption;
+    }
+
+    public function options(array $options = []): self
+    {
+        $this->options = $options;
+
+        return $this;
+    }
+
+    public function getOptions(): array
+    {
+        return $this->options;
+    }
+    
+    public function getKeys(): array
+    {
+        return collect($this->getOptions())
+            ->keys()
+            ->map(fn ($value) => (string) $value)
+            ->filter(fn ($value) => strlen($value))
+            ->values()
+            ->toArray();
+    }
+
+}
